+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.2] - 2018-03-10
+- Added: Modify an app entry
+- Added: Check if another instance is running
+- Slightly smaller font size in tray menu
+
 ## [1.1] - 2018-02-27
 - Launch apps with proper working directory